test/fixtures/wpt/wasm/webapi/esm-integration/worker.tentative.html

<!DOCTYPE html>
<title>Testing WebAssembly worker</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type=module>
setup({ single_test: true });
const worker = new Worker("resources/worker.wasm", { type: "module" });
worker.onmessage = (msg) => {
  assert_equals(msg, 42);
  done();
}
</script>
