LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/worker.tentative-expected.txt

CONSOLE MESSAGE: SyntaxError: Invalid character: '\0'

FAIL Testing WebAssembly worker Script error.
